requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
backports.functools-lru-cache==1.6.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chucknorris==2.8
cryptography==3.1.1
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
iocextract==1.13.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.5.3
python-telegram-bot==13.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
six==1.12.0
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.0.31
wrapt==1.12.1
yml==0.0.1
zipp==3.1.0
